requirements.txt
pyodbc
python-dotenv
requests
python-telegram-bot